Get Conservation District Boundary Geometry for a Washington County

Get Conservation District Boundary Geometry for a Washington County

Usage

get_cd_polygon(district_name, crs = 4326)

Arguments

district_name

Character. Name of the district (e.g., "Columbia" or c("Columbia","Palouse")

crs

Integer or character. Desired coordinate reference system (default: 4326)

Value

An sf object with the swcd boundary

Examples

get_cd_polygon(c("Columbia","Palouse"))
#> Simple feature collection with 2 features and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -118.2423 ymin: 45.99909 xmax: -117.0395 ymax: 47.06667
#> Geodetic CRS:  WGS 84
#> # A tibble: 2 × 3
#>   CNSVDST     swcd_name                                                 geometry
#> * <chr>       <chr>                                           <MULTIPOLYGON [°]>
#> 1 Columbia CD Columbia  (((-117.8521 46.62465, -117.8522 46.61909, -117.8522 46…
#> 2 Palouse CD  Palouse   (((-117.0661 47.06666, -117.0661 47.06654, -117.0659 47…

get_cd_polygon("Columbia")
#> Simple feature collection with 1 feature and 2 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -118.2423 ymin: 45.99909 xmax: -117.6033 ymax: 46.62465
#> Geodetic CRS:  WGS 84
#> # A tibble: 1 × 3
#>   CNSVDST     swcd_name                                                 geometry
#> * <chr>       <chr>                                           <MULTIPOLYGON [°]>
#> 1 Columbia CD Columbia  (((-117.8521 46.62465, -117.8522 46.61909, -117.8522 46…