Get List of Layers from Columbia GeoServer

This function queries a GeoServer WMS GetCapabilities endpoint and extracts a list of available WMS layer names. By default, it points to the Columbia County WA workspace.

Usage

show_geoserver_layers(
  geoserver_url =
    "https://geoserver.megaloptera-data.com/geoserver/Columbia/wms?service=WMS&version=1.1.0&request=GetCapabilities"
)

Arguments

capabilities_url

A character string specifying the WMS GetCapabilities URL. Defaults to the Columbia workspace on your GeoServer: "http://142.93.92.104:8080/geoserver/Columbia/wms?service=WMS&version=1.1.0&request=GetCapabilities"

Value

A character vector of layer names (e.g., "Columbia:usda_2024_cropland").

Examples

if (FALSE) { # \dontrun{

show_geoserver_layers()
} # }